requirements.txt
torch>=2.0
torch-geometric>=2.4
networkx>=3.0
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
joblib>=1.3
